+++ libraries/EpoxyMockFastLED/FastLED.cpp
@@ -1,0 +1,16 @@
+/*
+ * MIT License
+ * Copyright (c) 2013 FastLED
+ * Copyright (c) 2021 Brian T. Park
+ */
+
+#include "FastLED.h"
+
+CFastLED FastLED;
+
+const TProgmemRGBPalette16 PartyColors_p = {
+    0x5500AB, 0x84007C, 0xB5004B, 0xE5001B,
+    0xE81700, 0xB84700, 0xAB7700, 0xABAB00,
+    0xAB5500, 0xDD2200, 0xF2000E, 0xC2003E,
+    0x8F0071, 0x5F00A1, 0x2F00D0, 0x0007F9
+};